Updated DataSourcer and comments.
