+++ api/api.go
@@ -7,12 +7,12 @@ import (
 	"go-there/data"
 )
 
+// DataSourcer represents the datasource.DataSource methods needed by the api package to access the data.
 type DataSourcer interface {
 	SelectUser(username string) (data.User, error)
 	SelectUserLogin(username string) (data.User, error)
 	SelectApiKeyHashByUser(username string) ([]byte, error)
 	SelectUserLoginByApiKeySalt(apiKeySalt string) (data.User, error)
-	SelectApiKeyHashBySalt(apiKeySalt string) ([]byte, error)
 	InsertUser(user data.User) error
 	DeleteUser(username string) error
 	UpdatetUserPassword(user data.User) error
@@ -21,6 +21,7 @@ type DataSourcer interface {
 	DeletePath(path data.Path) error
 }
 
+// Init initializes the API paths from the provided configuration and add them to the *gin.Engine.
 func Init(conf *config.Configuration, e *gin.Engine, ds DataSourcer) {
 	ep := conf.Endpoints["manage_users"]
 	if ep.Enabled {
