+++ datasource/datasource.go
@@ -11,10 +11,13 @@ import (
 	"go-there/data"
 )
 
+// DataSource represents the source of the applciation's data.
 type DataSource struct {
 	db *sqlx.DB
 }
 
+// Init initializes and tries to connect to the database defined in the configuration. If it cannot connect, an error is
+// returned.
 func Init(config *config.Configuration) (*DataSource, error) {
 	var err error
 	ds := new(DataSource)
@@ -39,6 +42,7 @@ func Init(config *config.Configuration) (*DataSource, error) {
 	return ds, nil
 }
 
+// SelectUser fetches an complete user by his username in the database.
 func (ds *DataSource) SelectUser(username string) (data.User, error) {
 	u := data.User{}
 	err := ds.db.Get(&u, ds.db.Rebind("SELECT * FROM users WHERE username=?"), username)
@@ -50,6 +54,7 @@ func (ds *DataSource) SelectUser(username string) (data.User, error) {
 	return u, nil
 }
 
+// SelectUserLogin fetches the username,is_admin,password_hash as a user by his username in the database.
 func (ds *DataSource) SelectUserLogin(username string) (data.User, error) {
 	u := data.User{}
 	err := ds.db.Get(&u, ds.db.Rebind("SELECT username,is_admin,password_hash FROM users WHERE username=?"), username)
@@ -61,6 +66,7 @@ func (ds *DataSource) SelectUserLogin(username string) (data.User, error) {
 	return u, nil
 }
 
+// SelectApiKeyHashByUser fetches a full API key hash from the database by a username.
 func (ds *DataSource) SelectApiKeyHashByUser(username string) ([]byte, error) {
 	ak := make([]byte, 0)
 	err := ds.db.Get(&ak, ds.db.Rebind("SELECT api_key_hash FROM users WHERE username=?"), username)
@@ -72,6 +78,7 @@ func (ds *DataSource) SelectApiKeyHashByUser(username string) ([]byte, error) {
 	return ak, nil
 }
 
+// SelectUserLoginByApiKeySalt fetches the username,is_admin,api_key_hash of a user, by his API key salt.
 func (ds *DataSource) SelectUserLoginByApiKeySalt(apiKeySalt string) (data.User, error) {
 	u := data.User{}
 	err := ds.db.Get(&u, ds.db.Rebind("SELECT username,is_admin,api_key_hash FROM users WHERE api_key_salt=?"), apiKeySalt)
@@ -83,6 +90,7 @@ func (ds *DataSource) SelectUserLoginByApiKeySalt(apiKeySalt string) (data.User,
 	return u, nil
 }
 
+// SelectApiKeyHashBySalt fetches the full API key hash by its salt.
 func (ds *DataSource) SelectApiKeyHashBySalt(apiKeySalt string) ([]byte, error) {
 	ak := make([]byte, 0)
 	err := ds.db.Get(&ak, ds.db.Rebind("SELECT api_key_hash FROM users WHERE api_key_salt=?"), apiKeySalt)
@@ -94,6 +102,8 @@ func (ds *DataSource) SelectApiKeyHashBySalt(apiKeySalt string) ([]byte, error) 
 	return ak, nil
 }
 
+// InsertUser tries to add a new user to the database. If a user with the same name or API key salt exists,
+// data.ErrSqlDuplicateRow is returned.
 func (ds *DataSource) InsertUser(user data.User) error {
 	_, err := ds.db.NamedExec(
 		"INSERT INTO users (username,is_admin,password_hash,api_key_salt,api_key_hash) "+
@@ -108,24 +118,26 @@ func (ds *DataSource) InsertUser(user data.User) error {
 		} else {
 			return data.ErrSql
 		}
-
 	}
 
-	return err
+	return nil
 }
 
+// UpdatetUserPassword updates an user's password in the database.
 func (ds *DataSource) UpdatetUserPassword(user data.User) error {
 	_, err := ds.db.NamedExec("UPDATE users SET password_hash=:password_hash WHERE username=:username", user)
 
 	return err
 }
 
+// UpdatetUserPassword updates an user's API key in the database.
 func (ds *DataSource) UpdatetUserApiKey(user data.User) error {
 	_, err := ds.db.NamedExec("UPDATE users SET api_key_hash=:api_key_hash,api_key_salt=:api_key_salt WHERE username=:username", user)
 
 	return err
 }
 
+// DeleteUser deletes a user in the database by his userame.
 func (ds *DataSource) DeleteUser(username string) error {
 	_, err := ds.db.Exec(ds.db.Rebind("DELETE FROM users WHERE username=?"), username)
 
@@ -136,6 +148,7 @@ func (ds *DataSource) DeleteUser(username string) error {
 	return nil
 }
 
+// GetTarget gets a target in the database from a path.
 func (ds *DataSource) GetTarget(path string) (string, error) {
 	t := ""
 	err := ds.db.Get(&t, ds.db.Rebind("SELECT target FROM go WHERE path=?"), path)
@@ -152,6 +165,7 @@ func (ds *DataSource) GetTarget(path string) (string, error) {
 	return t, nil
 }
 
+// InsertPath adds a data.Path to the database.
 func (ds *DataSource) InsertPath(path data.Path) error {
 	_, err := ds.db.NamedExec("INSERT INTO go (path,target,user,is_public) VALUES (:path,:target,:user,:is_public)", path)
 
@@ -170,6 +184,7 @@ func (ds *DataSource) InsertPath(path data.Path) error {
 	return err
 }
 
+// InsertPath deletes a data.Path in the database.
 func (ds *DataSource) DeletePath(path data.Path) error {
 	_, err := ds.db.NamedExec("DELETE FROM go WHERE path=:path", path)
